save dashboard settings
